chore: various docs parsing fixes
Removes a remark plugin used to change `bash session` language
code blocks to `bash`. This plugin was originally implemented as
in https://github.com/electron/i18n/pull/903 as a fix for
unknown language errors.

Instead of parsing the generated HTML with Cheerio, we use
a remark plugin to navigate the Markdown AST in order to
fix the URLs of relative links. There should be no change
to the output, but this code should be cleaner and more
encapsulated.

Removes the `sections` property and instead exposes a document's
entire HTML as an `html` property. We originally added the sections
property in #222, but the feature was not very utilized and this
method of parsing led to issues when two section headings had the
same name (see #1713).

Close #1786